Revert "[policy] Don't spam resource requests to log."
This reverts commit 2a39afc295114f8ac7b9385f925594055eeaf031.

Looking at the prolog side, request variable is used when
constructing the resource sets. My guess is dres is cleaning
the variable unless the variable is referenced inside dres
side, and then prolog side doesn't see the variable and dies.
So this log line needs to exist for the whole to work.

[policy] Add echo back to resource_request. Contributes to MER#1143