[policy] Mute feedbacksounds for bluetooth accessories. Fixes JB#34794
Many bluetooth devices are so slow to start up that the first feedback
sound wouldn't be audible, and there might be some pop when headset
amplifier goes on. Then when bluetooth is suspended there may be another
pop when the amplifier shuts down. So in worst case scenario there is
just unnecessary pops twice per sound and feedback sound isn't even
audible.