[common] Workaround to reference variables. Fixes MER#1735
echo line was needed in the request because due to how dres vm is
implemented, unless variables are referenced in requests they aren't
visible anymore when prolog actions are handled in the request.

As a workaround create facts which can then be assigned with the
variables and no need to echo anymore.