[common] Never route to bthsp. JB#37753
Only route to hsp/hfp headsets when a call is ongoing. For the call
cases we have specific bthspforcall device, which is then used.

The change is made so that audio can be routed to sane-quality location
(for example ihf, wired headphones) even if the bluetooth headset is
connected and ready for call cases. It doesn't make sense to route audio
to bluetooth headset unless it has A2DP functionality.