Introduce preferred audio feature.
Preferred audio can be set only for devices that are
currently selectable.

When preferred audio is set and multiple audio devices
are valid for audio route, use the preferred info as
extra point when making routing decisions. If only one
device is preferred, always use that, if multiple devices
are preferred, use the device that would be used if only
the normal routing rules would be considered.

Example:
By default we route to headset when headset and ihf are
connected.
If ihf is set as preferred, we route to ihf instead.
If both headset and ihf are preferred, we route to headset.

[common] Introduce preferred audio feature. JB#41568