Do not force release resources if route change is requested.
When going from private route to public route normally player resources
are released to force playback pause when for example disconnecting
headset from device. But if the route change to public is initiated by
user (change route to speaker when listening to fm radio for instance)
do not force the release.

[common] Do not force release resources when route change is requested. JB#36733